# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puryaideu Api",
        "description": "Implementation of Swagger with in Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user_booking": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Book",
                "operationId": "book",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origin": {
                                        "type": "string"
                                    },
                                    "destination": {
                                        "type": "string"
                                    },
                                    "distance": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "string"
                                    },
                                    "passenger_number": {
                                        "type": "number"
                                    },
                                    "vehicle_type": {
                                        "type": "number"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "origin": "Kathmandu",
                                    "destination": "biratnagar",
                                    "distance": "10.5km",
                                    "duration": "2 hours",
                                    "passenger_number": "2",
                                    "vehicle_type": "2",
                                    "name": "Deepika",
                                    "phone_number": "98132313"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Booking",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statusB{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "User Booking Status",
                "operationId": "App\\Http\\Controllers\\Api\\BookingController::status",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statusCancel{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Cancel booking",
                "operationId": "App\\Http\\Controllers\\Api\\BookingController::cancel_ride",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statusOfRider{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Rider Booking Status",
                "operationId": "App\\Http\\Controllers\\Api\\BookingController::statusRider",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/estimated_price": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\BookingController::estimated_price",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/specific_user": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\LoginController::user_id",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User details displayed successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}